+++ src/hydrate/runner/index.ts
@@ -1,3 +1,3 @@
 export { createWindowFromHtml } from './create-window';
 export { hydrateDocument, renderToString, serializeDocumentToString, streamToString } from './render';
-export { deserializeProperty, serializeProperty } from './serialize';
+export { deserializeProperty, serializeProperty } from '@utils';
