+++ src/utils/index.ts
@@ -4,6 +4,7 @@ export * from './format-component-runtime-meta';
 export * from './helpers';
 export * from './is-glob';
 export * from './is-root-path';
+export * from './local-value';
 export * from './logger/logger-rollup';
 export * from './logger/logger-typescript';
 export * from './logger/logger-utils';
@@ -12,8 +13,11 @@ export * from './output-target';
 export * from './path';
 export * from './query-nonce-meta-tag-content';
 export * from './regular-expression';
+export * from './remote-value';
 export * as result from './result';
+export * from './serialize';
 export * from './sourcemaps';
+export * from './types';
 export * from './url-paths';
 export * from './util';
 export * from './validation';
